Inventory

+++ src/sql/Inventory Table.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS inventory (
 	id INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
 	store_id INT NOT NULL,
 	goods_id INT NOT NULL,
-	quantity DOUBLE,
+	quantity INT,
 	selling_price DOUBLE,
 	CONSTRAINT fk_store_id FOREIGN KEY (store_id) REFERENCES store(store_id) ON UPDATE CASCADE ON DELETE CASCADE,
 	CONSTRAINT fk_goods_id FOREIGN KEY (goods_id) REFERENCES goods(goods_id) ON UPDATE CASCADE ON DELETE CASCADE
